+++ test/fixtures/another-package/webpack.config.js
@@ -6,6 +6,7 @@ module.exports = {
   mode: 'development',
   entry: 'index.js',
   output: {
+    path: path.resolve(__dirname),
     filename: '[name].js'
   },
   module: {